backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
firebase-admin==6.4.0
Werkzeug==2.3.7
dnspython==2.4.2
gunicorn==21.2.0
bson==0.5.10